utils/model_manager.py
# 肾病AI模型管理工具类
from pathlib import Path
import yaml
import logging

logger = logging.getLogger(__name__)

class RenalAIModelManager:
    def __init__(self, config_path: Path):
        """初始化模型管理器，加载配置"""
        self.config_path = config_path
        self.config = self.load_config()
        # 占位：后续添加模型加载逻辑
        self.models = self.load_models()

    def load_config(self) -> dict:
        """加载模型配置文件"""
        with open(self.config_path, "r", encoding="utf-8") as f:
            return yaml.safe_load(f)

    def load_models(self) -> dict:
        """加载AI模型（占位，后续实现）"""
        logger.info("加载AI模型（占位，未加载实际模型）")
        return {
            "pathology_report": None,
            "glomeruli_count": None,
            "nuclei_count": None
        }

    def generate_pathology_report(self, file_path: Path, parameters: dict) -> dict:
        """生成病理报告（占位）"""
        logger.info(f"生成病理报告：{file_path.name}")
        return {
            "report_id": f"report_{Path(file_path).stem}",
            "file_name": file_path.name,
            "findings": "肾小球形态基本正常，未见明显病变，细胞核计数在正常范围",
            "conclusion": "未见明显肾病病理特征",
            "confidence": 0.92,
            "parameters": parameters
        }

    def count_glomeruli(self, file_path: Path, parameters: dict) -> dict:
        """肾小球计数（占位）"""
        logger.info(f"肾小球计数：{file_path.name}")
        return {
            "count": 28,  # 模拟计数结果
            "density": 1.2,  # 肾小球密度
            "file_name": file_path.name,
            "parameters": parameters
        }

    def count_nuclei_in_glomerulus(self, file_path: Path, parameters: dict) -> dict:
        """肾小球内细胞核计数（占位）"""
        logger.info(f"细胞核计数：{file_path.name}")
        return {
            "average_nuclei_per_glomerulus": 126,  # 模拟平均计数
            "total_nuclei": 3528,
            "file_name": file_path.name,
            "parameters": parameters
        }