utils/wsi_processor.py
# 病理全切片图像（WSI）处理工具类
from pathlib import Path
from PIL import Image
import numpy as np

class WSIProcessor:
    def is_wsi_file(self, filename: str) -> bool:
        """验证是否为支持的病理图像格式（占位，可扩展为svs/tiff等）"""
        supported_ext = [".svs", ".tiff", ".tif", ".png", ".jpg", ".jpeg"]
        return Path(filename).suffix.lower() in supported_ext

    def get_wsi_metadata(self, file_path: Path) -> dict:
        """获取病理图像元数据（占位）"""
        with Image.open(file_path) as img:
            return {
                "width": img.width,
                "height": img.height,
                "mode": img.mode,
                "format": img.format,
                "size_mb": round(file_path.stat().st_size / 1024 / 1024, 2)
            }

    def generate_thumbnail(self, file_path: Path, thumbnail_path: Path, size: tuple = (256, 256)) -> bool:
        """生成病理图像缩略图（占位）"""
        try:
            with Image.open(file_path) as img:
                img.thumbnail(size)
                img.save(thumbnail_path, "JPEG")
            return True
        except Exception as e:
            print(f"缩略图生成失败：{e}")
            return False